social_media_api/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.8
django-filter==24.3
django-heroku==0.3.1
django-restframework==0.0.1
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
mysql-connector-python==9.0.0
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
selenium==4.23.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
